Makefile for csv2sql
Makefile used to generate multi platform 32bit and 64bit versions of csv2sql using standard golang compiler. This is optional - just use:

    go build csv2sql.go 

to create the program on its own - but is useful if you want / need to generate multiple versions with one command, and you have make available on your computer of course.